+++ src/thor/__init__.py
@@ -12,9 +12,8 @@ from .biotsavart import (
 )
 from .materials import MU0, FreeSpace, LinearMaterial, NonlinearMaterial, BHCurve
 from . import mesh
-from . import test_utils
+from . import testing
 from .magnetization import mag_force
-# import .test_utils as test_utils
 
 
 __all__ = [
@@ -25,7 +24,7 @@ __all__ = [
     "BHCurve",
     "bfield_direct",
     "bfield_octree",
-    "test_utils",
+    "testing",
     "mesh",
     "bfield_dualtree",
     "bfield_hexahedron",
