+++ tests/benchmark.py
@@ -20,7 +20,7 @@ interactions = np.zeros(nbenches)
 
 for (i, mesh_size) in enumerate(mesh_sizes):
 
-    centroids, vol, jdensity = thor.test_utils.make_helmholtz(mesh_size)
+    centroids, vol, jdensity = thor.testing.make_helmholtz(mesh_size)
     n = centroids.shape[0] 
     interactions[i] = n*n
 
@@ -82,4 +82,3 @@ ax.set_yscale('log')
 ax.set_title(f"Thor Benchmarks: Helmholtz Coil Problem\n$\\theta={theta:.2}$")
 ax.legend()
 fig.savefig("tests/benchmarks_throughput.png")
-
