+++ tests/benchmark_error.py
@@ -1,4 +1,4 @@
-from thor.test_utils import make_helmholtz
+from thor.testing import make_helmholtz
 
 import thor 
 import numpy as np 
@@ -15,7 +15,7 @@ for (i, theta) in enumerate(theta_vals):
     bmag_direct = np.linalg.norm(bdirect, axis=1) 
     bmag_octree = np.linalg.norm(boctree, axis=1) 
     # errs[i] = thor.mean_relative_error(bmag_direct, bmag_octree)
-    errs[i] = thor.test_utils.smape(bmag_direct, bmag_octree)
+    errs[i] = thor.testing.smape(bmag_direct, bmag_octree)
 
 print(errs)
 fig, ax = plt.subplots()
