+++ tests/example.py
@@ -6,7 +6,7 @@ from matplotlib import pyplot as plt
 size = 1.0
 theta = 0.5 
 nthreads = 0
-centroids, vol, jdensity = thor.test_utils.make_helmholtz(size)
+centroids, vol, jdensity = thor.testing.make_helmholtz(size)
 
 # centroids = np.vstack((centroids, centroids))
 # vol = np.hstack((vol, vol))
