+++ tests/test_tet.py
@@ -59,9 +59,9 @@ bmag_direct_pt_axis = np.linalg.norm(bdirect_pt_axis, axis=1)
 bmag_direct_tet_axis = np.linalg.norm(bdirect_tet_axis, axis=1) 
 bmag_octree_pt_axis = np.linalg.norm(boctree_pt_axis, axis=1)
 bmag_octree_tet_axis = np.linalg.norm(boctree_tet_axis, axis=1)
-err_direct_pt_axis = thor.test_utils.smape(bmag_direct_tet_axis, bmag_direct_pt_axis)
-err_octree_pt_axis = thor.test_utils.smape(bmag_direct_tet_axis, bmag_octree_pt_axis)
-err_octree_tet_axis = thor.test_utils.smape(bmag_direct_tet_axis, bmag_octree_tet_axis)
+err_direct_pt_axis = thor.testing.smape(bmag_direct_tet_axis, bmag_direct_pt_axis)
+err_octree_pt_axis = thor.testing.smape(bmag_direct_tet_axis, bmag_octree_pt_axis)
+err_octree_tet_axis = thor.testing.smape(bmag_direct_tet_axis, bmag_octree_tet_axis)
 print("Mean error along solenoid axis (|z| < radius): ")
 print(f"\tdirect (pt):  {err_direct_pt_axis*100:.2}%")
 print(f"\toctree (pt):  {err_octree_pt_axis*100:.2}%")
@@ -100,9 +100,9 @@ bmag_direct_pt = np.linalg.norm(bdirect_pt, axis=1)
 bmag_direct_tet = np.linalg.norm(bdirect_tet, axis=1) 
 bmag_octree_pt = np.linalg.norm(boctree_pt, axis=1) 
 bmag_octree_tet = np.linalg.norm(boctree_tet, axis=1) 
-err_mesh_pt = thor.test_utils.smape(bmag_direct_tet, bmag_octree_pt)
-err_mesh_pt_direct = thor.test_utils.smape(bmag_direct_tet, bmag_direct_pt)
-err_mesh_tet = thor.test_utils.smape(bmag_direct_tet, bmag_octree_tet)
+err_mesh_pt = thor.testing.smape(bmag_direct_tet, bmag_octree_pt)
+err_mesh_pt_direct = thor.testing.smape(bmag_direct_tet, bmag_direct_pt)
+err_mesh_tet = thor.testing.smape(bmag_direct_tet, bmag_octree_tet)
 print(f"Mean fields error within the mesh (pt/direct sources):  {err_mesh_pt_direct*100:.2}%")
 print(f"Mean fields error within the mesh (pt/octree sources):  {err_mesh_pt*100:.2}%")
 print(f"Mean fields error within the mesh (tet/octree sources): {err_mesh_tet*100:.2}%")
